+++ src/main/java/com/github/iridatelegrambot/command/AddOrderCommand.java
@@ -1,6 +1,5 @@
 package com.github.iridatelegrambot.command;
 
-import com.github.iridatelegrambot.bot.CheckUpdateOnPost;
 import com.github.iridatelegrambot.service.SendMessageService;
 import com.github.iridatelegrambot.service.statuswait.WaitDocument;
 import org.telegram.telegrambots.meta.api.objects.Update;
@@ -10,12 +9,10 @@ import static com.github.iridatelegrambot.service.statuswait.WaitTypeStatus.ADD;
 public class AddOrderCommand implements Command{
 
     private final SendMessageService sendMessageService;
-    private final CheckUpdateOnPost checkUpdateOnPost;
 
     public static final String ADDORDER_MESSAGE = "Введите номер заказа на перемещение: ";
-    public AddOrderCommand(SendMessageService sendMessageService,CheckUpdateOnPost checkUpdateOnPost) {
+    public AddOrderCommand(SendMessageService sendMessageService) {
         this.sendMessageService = sendMessageService;
-        this.checkUpdateOnPost = checkUpdateOnPost;
     }
 
     @Override