+++ src/main/java/com/github/iridatelegrambot/command/CallbackCommand/OrderMenuCallbackCommand.java
@@ -1,6 +1,5 @@
 package com.github.iridatelegrambot.command.CallbackCommand;
 
-import com.github.iridatelegrambot.bot.CheckUpdateOnPost;
 import com.github.iridatelegrambot.command.AddInvoiceCommand;
 import com.github.iridatelegrambot.entity.Order;
 import com.github.iridatelegrambot.service.OrderService;
@@ -14,13 +13,10 @@ import java.util.Optional;
 public class OrderMenuCallbackCommand implements CallbackCommand {
     private final OrderService orderService;
     private final SendMessageService sendMessageService;
-    private final CheckUpdateOnPost checkUpdateOnPost;
 
-    public OrderMenuCallbackCommand(OrderService orderService, SendMessageService sendMessageService,
-                                    CheckUpdateOnPost checkUpdateOnPost) {
+    public OrderMenuCallbackCommand(OrderService orderService, SendMessageService sendMessageService) {
         this.orderService = orderService;
         this.sendMessageService = sendMessageService;
-        this.checkUpdateOnPost = checkUpdateOnPost;
     }
 
     @Override
@@ -42,9 +38,8 @@ public class OrderMenuCallbackCommand implements CallbackCommand {
             Message message = callbackQuery.getMessage();
             message.setText("/add_invoice");
             update.setMessage(message);
-            AddInvoiceCommand addInvoiceCommand = new AddInvoiceCommand(sendMessageService,checkUpdateOnPost);
+            AddInvoiceCommand addInvoiceCommand = new AddInvoiceCommand(sendMessageService);
             addInvoiceCommand.execute(update);
-            checkUpdateOnPost.setOrderToBond(chatId,idOrder);
         } else if (subtype.equals("delete")){
             orderService.delete(idOrder);
             sendMessageService.sendMessage(chatId.toString(),"Заказ с номером \"" + order.getNumber() +