+++ src/main/java/com/github/iridatelegrambot/command/CommandName.java
@@ -9,8 +9,6 @@ public enum CommandName {
     ADDORDER("/add_order"),
     ADDINVOICE("/add_invoice"),
     SHOWACTIVEORDER("/show_active_orders"),
-    SHOWALLORDER("/show_all_order"),
-    SHOWALLINVOICE("/show_all_invoice"),
     NO("nocommand");
 
     private final String commandName;