+++ src/main/java/com/github/iridatelegrambot/repository/ConditionBotRepository.java
@@ -1,7 +1,0 @@
-package com.github.iridatelegrambot.repository;
-
-import com.github.iridatelegrambot.entity.ConditionBot;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ConditionBotRepository extends JpaRepository<ConditionBot, Integer> {
-}