+++ src/main/resources/db/migration/V0003__add_invoices_table.sql
@@ -5,6 +5,11 @@ ALTER TABLE orders ADD date_created TIMESTAMP NULL DEFAULT CURRENT_TIMESTAMP;
 DROP TABLE IF EXISTS invoices;
 
 
+ALTER TABLE user_telegram ADD user_name varchar (100);
+ALTER TABLE user_telegram ADD first_name varchar (100);
+ALTER TABLE user_telegram ADD admin boolean DEFAULT false;
+
+
 CREATE TABLE invoices (
     id_invoice int primary key auto_increment,
     number_invoice varchar(30),