+++ src/test/java/com/github/iridatelegrambot/test/command/AbstractCommandTest.java
@@ -1,6 +1,5 @@
 package com.github.iridatelegrambot.test.command;
 
-import com.github.iridatelegrambot.bot.CheckUpdateOnPost;
 import com.github.iridatelegrambot.command.Command;
 import com.github.iridatelegrambot.service.SendMessageServiceImpl;
 import com.github.iridatelegrambot.service.UserTelegramService;
@@ -8,7 +7,6 @@ import com.github.iridatelegrambot.bot.IridaBot;
 import com.github.iridatelegrambot.service.buttons.InlineKeyboardService;
 import com.github.iridatelegrambot.service.buttons.MenuButtonsService;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 import org.telegram.telegrambots.meta.api.objects.Chat;
@@ -20,9 +18,8 @@ abstract class AbstractCommandTest {
     protected UserTelegramService mUserTelegramService = Mockito.mock(UserTelegramService.class);
     protected IridaBot mIridaBot = Mockito.mock(IridaBot.class);
     protected InlineKeyboardService mInlineKeyboardService = Mockito.mock(InlineKeyboardService.class);
-    protected CheckUpdateOnPost mCheckUpdateOnPost = Mockito.mock(CheckUpdateOnPost.class);
     protected MenuButtonsService menuButtonsService = Mockito.mock(MenuButtonsService.class);
-    protected SendMessageServiceImpl sendMessageService = new SendMessageServiceImpl(mIridaBot,mInlineKeyboardService,mCheckUpdateOnPost,menuButtonsService);
+    protected SendMessageServiceImpl sendMessageService = new SendMessageServiceImpl(mIridaBot,mInlineKeyboardService,menuButtonsService);
 
     abstract String getCommandName();