+++ src/test/resources/sql/condition_bot.sql
@@ -1,2 +1,0 @@
-INSERT INTO user_telegram VALUES ("100000001", 0 ,"bu1","bur1",0);
-INSERT INTO condition_bot VALUES (1,"100000001", 0, 0);