+++ src/test/resources/sql/delete_table.sql
@@ -1,4 +1,3 @@
-DELETE FROM condition_bot;
 DELETE FROM invoices;
 DELETE FROM orders;
 DELETE FROM user_telegram;