+++ RELEASE_NOTES.md
@@ -1,5 +1,11 @@
 # Release Notes
 
+## 0.5.0-SNAPSHOT
+
+*   IRIDA-6: Added the ability to show all open orders for the user.
+*   IRIDA-7: Added the ability to show the main menu with the main command.
+*   IRIDA-8: Added the menu to each active order, stat menu with the ability to show all documents and delete documents.
+
 ## 0.4.0-SNAPSHOT
 
 *   IRIDA-4: Added command /addorder and buttons markup for it.